requirements.txt
pyserial
pyyaml
configparser
pyinstaller
